backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.26.0
python-dotenv==1.0.1
google-generativeai==0.4.1
pydantic==2.6.1
